experiments/deepgram_platform/pyproject.toml
[project]
name = "deepgram-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio==5.0.0",
    "deepgram-sdk>=3.1.0",
    "python-dotenv==1.0.0",
    "openai>=1.0.0",
    "langchain-openai>=0.3.6",
    "langchain>=0.3.18",
    "langchain-community>=0.3.17",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
